功能 回复异常处理 重构 user,question字段类型为Long

+++ pjf/src/main/java/com/pjf/pjf/controller/PublishController.java
@@ -39,7 +39,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String question(@PathVariable(name = "id") Integer id, Model model) {
+    public String question(@PathVariable(name = "id") Long id, Model model) {
         QuestionDTO questionDTO = questionService.getById(id);
         model.addAttribute("title", questionDTO.getTitle());
         model.addAttribute("description", questionDTO.getDescription());
@@ -58,7 +58,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             Model model,
             HttpServletRequest request
     ) {
